etl-divisas/requirements.txt
requests==2.31.0
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.0
matplotlib==3.8.0
openpyxl==3.1.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23